adjust

+++ src/main/java/com/haoyong/preview/service/impl/AsyncServiceImpl.java
@@ -1,18 +1,15 @@
 package com.haoyong.preview.service.impl;
 
 import com.haoyong.preview.service.AsyncService;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.encryption.InvalidPasswordException;
-import org.jodconverter.DocumentConverter;
-import org.jodconverter.office.OfficeException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
+
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.scheduling.annotation.AsyncResult;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
 import java.io.File;
 import java.io.IOException;
 import java.util.concurrent.Future;
@@ -25,29 +22,39 @@ import java.util.concurrent.Future;
  * @create: 2020-12-30 11:50
  **/
 @Service
+@Slf4j
 public class AsyncServiceImpl implements AsyncService {
-    @Resource
-    DocumentConverter documentConverter;
+
 
     @Override
     @Async("taskExecutor")
     public Future<Integer> fileToPdf(File file,String targpath) {
         try {
             File file1 = new File(targpath);
-            documentConverter.convert(file)
-                    .to(file1)
-                    .execute();
+
             PDDocument load = PDDocument.load(file1);
             int numberOfPages = load.getNumberOfPages();
-            file.delete();
+            if (file.exists()) {
+                file.delete();
+            }
+
             return new AsyncResult<>(numberOfPages);
-        } catch (OfficeException e) {
-            e.printStackTrace();
         } catch (InvalidPasswordException e) {
             e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
         }
         return new AsyncResult<>(0);
+    }
+
+    @Override
+    @Async("taskExecutor")
+    public void deleteTemporaryFile(String targpath) {
+        File file = new File(targpath);
+
+
+        if (file.exists()) {
+            file.delete();
+        }
     }
 }
